# src/main/resources/static/swagger.yaml
swagger: "2.0"
info:
  description: "This is a description for toys exchange server."
  version: "1.0.0"
  title: "Swagger ToysExchange"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
host: "localhost:8080"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
  /toys:
    get:
      tags:
      - "toys"
      summary: "get the available toys"
      description: "Get the avaialable toys : not expired, not deleted and not suspended"
      operationId: "getToys"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "userId"
        type: "string"
        description: "the  userId of the user who requested the list of toys"
        required: true
      responses:
        200:
          description : "the list of available toys"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ToyVO"
        405:
          description: "Invalid input"
      security:
      - toystore_auth:
        - "write:toys"
        - "read:toys"
...
  /toys/{searchCriteria}:
    get:
      tags:
      - "toys"
      summary: "Finds toys corresponding to mulitple criterias"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findToys"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "searchCriteria"
        in: "path"
        description: "toys search criteria"
        required: true
        type: "array"
        items:
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ToyVO"
        400:
          description: "Invalid status value"
      security:
      - toystore_auth:
        - "write:toys"
        - "read:toys"
...
  /toys/order:
    post:
      tags:
      - "exchange"
      summary: "Place an order for a toy"
      description: ""
      operationId: "placeOrder"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "order placed for purchasing the pet"
        required: true
        schema:
          $ref: "#/definitions/Order"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        400:
          description: "Invalid Order"
  /toys/order/{orderId}:
    get:
      tags:
      - "exchange"
      summary: "Find exchange order by ID"
      description: "For valid response try integer IDs with value >= 1 and <= 10.         Other values will generated exceptions"
      operationId: "getOrderById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "orderId"
        in: "path"
        description: "ID of pet that needs to be fetched"
        required: true
        type: "integer"
        maximum: 10.0
        minimum: 1.0
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Order not found"
    delete:
      tags:
      - "exchange"
      summary: "Delete exchange order by ID"
      description: "For valid response try integer IDs with positive integer value.         Negative or non-integer values will generate API errors"
      operationId: "deleteOrder"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "orderId"
        in: "path"
        description: "ID of the order that needs to be deleted"
        required: true
        type: "integer"
        minimum: 1.0
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Order not found"
...
securityDefinitions:
  toystore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:toys: "modify pets in your account"
      read:toys: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Order:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      petId:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int32"
      shipDate:
        type: "string"
        format: "date-time"
      status:
        type: "string"
        description: "Order Status"
        enum:
        - "placed"
        - "approved"
        - "delivered"
      complete:
        type: "boolean"
        default: false
    xml:
      name: "Order"
...